refactor(http): extract shared retrying JSON POST sink
The CDR webhook sink (siphon-ai-cdr) and the lifecycle webhook sink
(siphon-ai-webhooks) each carried their own near-identical copy of the
retrying JSON-over-HTTP POST logic: config defaults, reqwest::Client
setup, the post_with_retry loop, transient-status classification,
exponential backoff, and the test scaffolding for all of it. A change
to delivery behavior — or the future HMAC signing called out in
DEV_PLAN §11.6 — would have to be made twice and kept in sync by hand.

Add a new `siphon-ai-http` crate exposing `RetryingPoster`: a
connection-pooling client that POSTs a `Serialize` payload and retries
transient failures (5xx/408/429) with exponential backoff. Both sinks
become thin adapters that build a `RetryingPoster` and spawn
`poster.post(..)`; the retry mechanics and their tests now live once.

Neither cdr nor webhooks references `reqwest` directly anymore, so the
direct dependency is dropped from both (`RetryingPoster::new` returns
`siphon_ai_http::BuildError`). webhooks' `serde_json` moves to
dev-dependencies (it was only ever used by tests).
