+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/cdr",
     "crates/config",
     "crates/core",
+    "crates/http",
     "crates/media-glue",
     "crates/routes",
     "crates/sip-glue",
@@ -28,6 +29,7 @@ siphon-ai-bridge      = { path = "crates/bridge" }
 siphon-ai-cdr         = { path = "crates/cdr" }
 siphon-ai-config      = { path = "crates/config" }
 siphon-ai-core        = { path = "crates/core" }
+siphon-ai-http        = { path = "crates/http" }
 siphon-ai-media-glue  = { path = "crates/media-glue" }
 siphon-ai-routes      = { path = "crates/routes" }
 siphon-ai-sip-glue    = { path = "crates/sip-glue" }
